Simplificado o contexto de dados da cana e adicionado mais um conversor, este transforma IEnumerable em um ObservableCollection.

+++ NFeFacil/ModeloXML/PartesProcesso/PartesNFe/EnderecoBase.cs
@@ -13,7 +13,7 @@ namespace NFeFacil.ModeloXML.PartesProcesso.PartesNFe
         [XmlElement("xBairro")]
         public string Bairro { get; set; }
         [XmlElement("cMun")]
-        public long CodigoMunicipio { get; set; }
+        public int CodigoMunicipio { get; set; }
         [XmlElement("xMun")]
         public string NomeMunicipio { get; set; }
         [XmlElement("UF")]
